first rough idea sketch
- currently just a JUnit Runner
- fire up all test cases within Futures and feed reports back via a
  LinkedBlockingQueue (not sure how JUnit’s reporting copes with
  concurrent notifications)
- using scala-async for writing the test cases

The obvious downside is that the reporting of the error will in most
cases not be able to point to the right source line, not sure what can
be done to improve this. In case of a general timeout this will
certainly not be possible (i.e. we cannot find out where it is currently
hanging).